pyproject.toml
[tool.poetry]
name = "miiiii"
version = "0.0.0"
description = "Mechanistic interpretability of irriducible integer identifiers"
authors = ["Noah Syrkis <desk@syrkis.com>"]
readme = "README.md"

[tool.poetry.scripts]
package_command = "main:main"

[tool.poetry.dependencies]
python = "^3.11"
oeis = "^2023.3.10"
tqdm = "^4.66.4"
requests = "^2.32.2"
einops = "^0.8.0"
matplotlib = "^3.9.0"
streamlit = "^1.35.0"          # remove maybe
pyyaml = "^6.0.1"
numpy-hilbert-curve = "^1.0.1" # remove probably
seaborn = "^0.13.2"
scikit-learn = "^1.5.0"
jupyterlab = "^4.2.2"
jax = "^0.4.31"
optax = "^0.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=miiiii --cov-report=term-missing"
testpaths = ["assert"]
